requirements.txt
amqp==2.2.2
appdirs==1.4.3
billiard==3.5.0.3
celery==4.1.0
contextlib2==0.5.5
cookies==2.2.1
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.5.0
dj-database-url==0.5.0
django-haystack==2.8.1
Django==2.1.5
django-allauth==0.35.0
django-cors-headers==2.2.0
django-rest-auth==0.9.3
django-widget-tweaks==1.4.3
djangorestframework==3.9.3
djangorestframework-expander==0.2.3
djangorestframework-gis==0.14
djangorestframework-jwt==1.11.0
git+https://git@github.com/chibisov/drf-extensions.git@master
drf-nested-routers==0.90.2
easy-thumbnails==2.5
et-xmlfile==1.0.1
factory-boy==2.7.0
fake-factory==0.6.0
fuzzywuzzy==0.16.0
geopy==1.13.0
gunicorn==19.8.1
inflection==0.3.1
itypes==1.1.0
jdcal==1.4
Jinja2==2.10
kombu==4.1.0
MarkupSafe==1.0
oauthlib==2.0.7
olefile==0.45.1
openpyxl==2.5.3
packaging==17.1
parse==1.8.2
parse-type==0.3.4
pbr==4.0.2
Pillow==4.0.0
psycopg2==2.7.4
PyJWT==1.6.1
pyparsing==2.2.0
python-dateutil==2.7.2
python-Levenshtein==0.12.0
python3-openid==3.1.0
pytz==2018.4
raven==6.7.0
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.8.0
responses==0.9.0
six==1.11.0
sqlparse==0.2.4
uritemplate==3.0.0
urllib3==1.22
uwsgi
vine==1.1.4
whitenoise==3.3.0
git+https://github.com/dtougas/xapian-haystack@master
django-webpack-loader==0.6.0
django-filter==2.1.0
